backend/pyproject.toml
[tool.poetry]
name = "codepulse-backend"
version = "0.1.0"
description = "CodePulse API - a code review platform"
authors = ["Houman Feizabadi <hamedfz19@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.3"
uvicorn = {version = "^0.43.0", extras = ["standard"]}
sqlalchemy = {version = "^2.0.49", extras = ["asyncio"]}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"
PyJWT = {version = "^2.12.1", extras = ["crypto"]}
bcrypt = "^5.0.0"
python-multipart = "^0.0.22"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
aiosqlite = "^0.22.1"
httpx = "^0.28.1"
ruff = "^0.15.9"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"